Move devices to root
 * Move .v1.devices to root
 * Update README on correcting device information

Sem-Ver: bugfix
