+++ src/aiohubspace/v1/controllers/fan.py
@@ -1,11 +1,11 @@
 """Controller holding and managing Hubspace resources of type `fan`."""
 
-from .. import device
-from ..device import HubspaceDevice
+from ... import device
+from ...device import HubspaceDevice
+from ...util import ordered_list_item_to_percentage
 from ..models import features
 from ..models.fan import Fan, FanPut
 from ..models.resource import DeviceInformation, ResourceTypes
-from ...util import ordered_list_item_to_percentage
 from .base import BaseResourcesController
 
 KNOWN_PRESETS = {"comfort-breeze"}
