+++ tests/v1/controllers/test_valve.py
@@ -4,9 +4,9 @@ import asyncio
 
 import pytest
 
+from aiohubspace.device import HubspaceState
 from aiohubspace.v1.controllers import event
 from aiohubspace.v1.controllers.valve import ValveController, features
-from aiohubspace.v1.device import HubspaceState
 
 from .. import utils
